+++ packages/mcp/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @faapi/mcp
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @faapi/faapi@1.0.2
+
 ## 1.0.1
 
 ### Patch Changes